authentication app restructure
