+++ authentication/models.py
@@ -16,4 +16,3 @@ class User(AbstractUser):
 
     def __str__(self):
         return f"{self.first_name} {self.last_name}"
-
