+++ authentication/serializer.py
@@ -1,41 +1,56 @@
+# from rest_framework import serializers
+# from rest_framework.validators import UniqueValidator
+# from rest_framework_simplejwt.tokens import Token
+# from authentication.models import User
+# from django.contrib.auth.password_validation import validate_password
+# from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
+
+
+from dj_rest_auth.registration.serializers import RegisterSerializer
 from rest_framework import serializers
-from rest_framework.validators import UniqueValidator
-from rest_framework_simplejwt.tokens import Token
-from authentication.models import User
-from django.contrib.auth.password_validation import validate_password
-from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
 
 
-class SignUpSerializer(serializers.ModelSerializer):
-    email = serializers.EmailField(required=True, validators=[UniqueValidator(queryset=User.objects.all())])
-    password = serializers.CharField(write_only=True, required=True, validators=[validate_password])
-    password2 = serializers.CharField(write_only=True, required=True)
+class CustomRegisterSerializer(RegisterSerializer):
+    first_name = serializers.CharField(max_length=50)
+    last_name = serializers.CharField(max_length=50)
 
-    class Meta:
-        model = User
-        fields = ['email', 'first_name', 'last_name', 'password','password2', 'is_owner']
+    def get_cleaned_data(self):
+        data = super().get_cleaned_data()
+        data['first_name'] = self.validated_data.get('first_name', '')
+        data['last_name'] = self.validated_data.get('last_name', '')
+        return data
 
-    def validate(self, attrs):
-        if attrs['password'] != attrs['password2']:
-            raise serializers.ValidationError({'password':'password fields did not match'})
-        return attrs
-    def create(self, validated_data):
-        user = User.objects.create(
-            email=validated_data['email'],
-            first_name=validated_data['first_name'],
-            last_name=validated_data['last_name'],
-            is_owner = validated_data['is_owner']
-        )
-        user.set_password(validated_data['password'])
-        user.save()
-        return user
+
+# class SignUpSerializer(serializers.ModelSerializer):
+#     email = serializers.EmailField(required=True, validators=[UniqueValidator(queryset=User.objects.all())])
+#     password = serializers.CharField(write_only=True, required=True, validators=[validate_password])
+#     password2 = serializers.CharField(write_only=True, required=True)
+
+#     class Meta:
+#         model = User
+#         fields = ['email', 'first_name', 'last_name', 'password','password2', 'is_owner']
+
+#     def validate(self, attrs):
+#         if attrs['password'] != attrs['password2']:
+#             raise serializers.ValidationError({'password':'password fields did not match'})
+#         return attrs
+#     def create(self, validated_data):
+#         user = User.objects.create(
+#             email=validated_data['email'],
+#             first_name=validated_data['first_name'],
+#             last_name=validated_data['last_name'],
+#             is_owner = validated_data['is_owner']
+#         )
+#         user.set_password(validated_data['password'])
+#         user.save()
+#         return user
     
 
 
-class JwtSerializer(TokenObtainPairSerializer):
+# class JwtSerializer(TokenObtainPairSerializer):
 
-    @classmethod
-    def get_token(cls, user) -> Token:
-        token = super().get_token(user)
-        token['email'] = user.email
-        return token
+#     @classmethod
+#     def get_token(cls, user) -> Token:
+#         token = super().get_token(user)
+#         token['email'] = user.email
+#         return token
