--- authentication/serializers/__init__.py
@@ -1,1 +1,0 @@
-from .register_serializer import CustomRegisterSerializer
