--- authentication/serializers/register_serializer.py
@@ -1,13 +1,0 @@
-from dj_rest_auth.registration.serializers import RegisterSerializer
-from rest_framework import serializers
-
-
-class CustomRegisterSerializer(RegisterSerializer):
-    first_name = serializers.CharField(max_length=50)
-    last_name = serializers.CharField(max_length=50)
-
-    def get_cleaned_data(self):
-        data = super().get_cleaned_data()
-        data['first_name'] = self.validated_data.get('first_name', '')
-        data['last_name'] = self.validated_data.get('last_name', '')
-        return data
