--- authentication/utils/helpers.py
@@ -1,4 +1,0 @@
-import re
-def validate_password(password):
-    password_pattern = "^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{8,}$"
-    return re.match(password_pattern, password)
